Add files via upload
String is a sequence of characters .But in Java, string is an object that represents a sequence of characters. In java, string objects are immutable. Immutable simply means unmodifiable or unchangeable. Once string object is created its data or state can't be changed but a new string object is created.

String Comparison :
1) equals( ) method
2) == Operator
3) compareTo( ) method